feat: router refactor

+++ src/app/home/events/event-edit/event-edit.component.ts
@@ -1,11 +1,8 @@
 import {AsyncPipe, NgIf} from '@angular/common';
 import {ChangeDetectionStrategy, Component, inject} from '@angular/core';
 
-import {combineLatest, filter, map} from 'rxjs';
-
 import {NgbInputDatepicker, NgbNavModule} from '@ng-bootstrap/ng-bootstrap';
 
-import {notNullAndUndefined} from 'dfts-helper';
 import {BiComponent} from 'dfx-bootstrap-icons';
 import {DfxTr} from 'dfx-translate';
 
@@ -14,8 +11,9 @@ import {AppSelectableBtnComponent} from '../../../_shared/ui/button/app-selectab
 import {AbstractModelEditComponent} from '../../../_shared/ui/form/abstract-model-edit.component';
 import {AppFormModule} from '../../../_shared/ui/form/app-form.module';
 import {CreateEventOrLocationDto, GetEventOrLocationResponse, UpdateEventOrLocationDto} from '../../../_shared/waiterrobot-backend';
-import {OrganisationsService} from '../../organisations/_services/organisations.service';
+import {SelectedOrganisationService} from '../../organisations/_services/selected-organisation.service';
 import {EventsService} from '../_services/events.service';
+import {SelectedEventService} from '../_services/selected-event.service';
 import {AppEventEditFormComponent} from './event-edit-form.component';
 
 @Component({
@@ -30,14 +28,21 @@ import {AppEventEditFormComponent} from './event-edit-form.component';
         <app-model-edit-save-btn (submit)="form?.submit()" [valid]="valid()" [editing]="entity !== 'CREATE'" />
 
         <ng-container *isEditing="entity">
-          <div *ngIf="(myUser$ | async)?.isAdmin">
-            <button class="btn btn-sm btn-outline-danger" (click)="onDelete(entity.id)">
-              <bi name="trash" />
-              {{ 'DELETE' | tr }}
-            </button>
-          </div>
+          @if (myUser()?.isAdmin) {
+            <div>
+              <button class="btn btn-sm btn-outline-danger" (click)="onDelete(entity.id)">
+                <bi name="trash" />
+                {{ 'DELETE' | tr }}
+              </button>
+            </div>
+          }
           <div>
-            <selectable-button class="my-2" [entity]="entity" [selectedEntityService]="eventsService" placement="top" />
+            <selectable-button
+              class="my-2"
+              [entityId]="entity.id"
+              [selectedId]="selectedEventService.selectedId()"
+              (selectedChange)="selectedEventService.setSelected($event)"
+            />
           </div>
         </ng-container>
       </scrollable-toolbar>
@@ -47,13 +52,12 @@ import {AppEventEditFormComponent} from './event-edit-form.component';
           <a ngbNavLink>{{ 'DATA' | tr }}</a>
           <ng-template ngbNavContent>
             <app-event-edit-form
-              *ngIf="vm$ | async as vm"
               #form
               (formValid)="setValid($event)"
               (submitUpdate)="submit('UPDATE', $event)"
               (submitCreate)="submit('CREATE', $event)"
-              [selectedOrganisationId]="vm.selectedOrganisation.id"
-              [formDisabled]="!vm.myUser.isAdmin"
+              [selectedOrganisationId]="selectedOrganisationId()!"
+              [formDisabled]="!myUser()!.isAdmin"
               [event]="entity"
             />
           </ng-template>
@@ -92,13 +96,9 @@ export class EventEditComponent extends AbstractModelEditComponent<
 
   defaultTab = 'DATA' as const;
 
-  myUser$ = inject(MyUserService).getUser$();
-
-  selectedOrganisation = inject(OrganisationsService).getSelected$;
-
-  vm$ = combineLatest([this.myUser$, this.selectedOrganisation.pipe(filter(notNullAndUndefined))]).pipe(
-    map(([myUser, selectedOrganisation]) => ({myUser, selectedOrganisation})),
-  );
+  myUser = inject(MyUserService).user;
+  selectedEventService = inject(SelectedEventService);
+  selectedOrganisationId = inject(SelectedOrganisationService).selectedId;
 
   constructor(public eventsService: EventsService) {
     super(eventsService);
